refactor(ampli): enforce initial user prop calls precedence
