+++ src/components/AmpliTrace/index.tsx
@@ -1,0 +1,1 @@
+export { AmpliTrace } from "./AmpliTrace";
