--- src/components/RouteTrace/RouteTrace.tsx
@@ -1,8 +1,0 @@
-import { useRouteTrace } from "./useRouteTrace";
-
-const RouteTrace = () => {
-  useRouteTrace();
-  return <></>;
-};
-
-export default RouteTrace;
