--- src/components/RouteTrace/index.ts
@@ -1,1 +1,0 @@
-export * from "./RouteTrace";
