--- src/components/WalletBalanceTrace/WalletBalanceTrace.tsx
@@ -1,8 +1,0 @@
-import { useWalletBalanceTrace } from "./useWalletBalanceTrace";
-
-const WalletBalanceTrace = () => {
-  useWalletBalanceTrace();
-  return <></>;
-};
-
-export default WalletBalanceTrace;
