--- src/components/WalletTrace/WalletTrace.tsx
@@ -1,6 +1,0 @@
-import { useWalletTrace } from "./useWalletTrace";
-
-export function WalletTrace() {
-  useWalletTrace();
-  return null;
-}
