--- src/components/WalletTrace/index.ts
@@ -1,1 +1,0 @@
-export { WalletTrace as default } from "./WalletTrace";
