+++ src/hooks/index.ts
@@ -14,3 +14,6 @@ export * from "./useConnection";
 export * from "./useIsWrongNetwork";
 export * from "./useStakingPool";
 export * from "./useApprove";
+export * from "./useWalletBalanceTrace";
+export * from "./useRouteTrace";
+export * from "./useWalletTrace";
